Move test suite generator to separate file
